+++ src/Components/NewsGrid/NewsGridItems.js
@@ -1,5 +1,5 @@
-import Component from "../core/Component.js";
-import { constants } from "../Data/constants.js";
+import Component from "../../core/Component.js";
+import { constants } from "../../Data/constants.js";
 
 export default class NewsGridItems extends Component {
   template() {
