+++ src/Components/NewsList/NewsList.js
@@ -1,5 +1,5 @@
-import { constants } from "../Data/constants.js";
-import Component from "../core/Component.js";
+import { constants } from "../../Data/constants.js";
+import Component from "../../core/Component.js";
 import PressNews from "./PressNews.js";
 
 export default class NewsList extends Component {
